moving files to correct folders
