+++ src/App.css
@@ -39,6 +39,7 @@ body {
 
 ::-webkit-scrollbar-track {
   background: rgba(0, 0, 0, 0.5);
+  border-radius: 10px;
 }
 
 ::-webkit-scrollbar-thumb {
