+++ src/Components/contentBuild.js
