+++ src/Components/contentStyle.js
